cgen: Exclude bits which are beyond the current end bit
If you have variable length instructions then decode-get-best-bits
could choose any bits as `best', even if they were not valid bits
for the instructions which it is trying to differentiate. When
this happened an assertion would be triggered when trying to
probe a bit beyond the mask of an instruction which was too short.

cgen/ChangeLog:

	* decode.scm (decode-get-best-bits): Remove any bits beyond
	startbit+decode-bitsize from the bits to be considered as 'best'.