cgen: Fix decoder switch when fetching more bits
When the bits decoder so far are not sufficient to differentiate
the instructions to be decoder more bits must be fetched. Previously
when these bits were fetched they overwrote the existing
instruction value, however the generated switch statement is written
to check bits as though the extra bits were appended to the existing
instruction.

cgen/ChangeLog:

	* utils-sim.scm (gen-decoder-switch): When fetching more bits,
	shift the bits into the correct offset (<< 'startbit') in the
	instruction value.